pyproject.toml
[tool.poetry]
name = "fern_flask"
version = "0.1.0"
description = "A knowledge capture tool for data hoarders"
authors = ["csmizzle <chrissmith700@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
Flask-Cors = "^3.0.10"
Flask-JWT = "asdf"
"pdfminer.six" = "^20211012"
python-docx = "^0.8.11"
spacy = "^3.2.0"
nltk = "^3.6.5"
Flask-PyMongo = "^2.3.0"
mongoengine = "^0.23.1"
graphene-mongo = "^0.2.13"
Flask-GraphQL = "^2.0.1"
python-magic = "^0.4.24"
mongomock = "^3.23.0"
bs4 = "^0.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
